backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.0
attrs==25.4.0
azure-core==1.39.0
azure-storage-blob==12.28.0
backports.zstd==1.3.0
bcrypt==5.0.0
black==25.9.0
boto3==1.40.39
botocore==1.40.39
brotli==1.2.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.1
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.1
et_xmlfile==2.0.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.3
firebase_admin==7.1.0
flake8==7.3.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-genai==1.60.0
google-generativeai==0.8.6
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.6.4
httpx==0.28.1
huggingface_hub==1.3.3
hyperframe==6.1.0
icalendar==5.0.13
idna==3.10
importlib_metadata==8.7.1
inflate64==1.0.4
iniconfig==2.1.0
invoke==2.2.1
isodate==0.7.2
isort==6.0.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jq==1.10.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
msgpack==1.1.2
multidict==6.7.0
multivolumefile==0.2.3
mypy==1.18.2
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.99.9
openpyxl==3.1.2
packaging==25.0
pandas==2.3.2
paramiko==4.0.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.1
py-vapid==1.9.2
py7zr==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.7
pycodestyle==2.14.0
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydyf==0.12.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyNaCl==1.6.2
pyparsing==3.2.5
pyphen==0.17.2
pyppmd==1.3.1
pytest==8.4.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
pywebpush==2.1.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.19.0
rich==14.1.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
scikit-learn==1.8.0
scipy==1.17.0
sendgrid==6.12.5
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
staticmap==0.5.7
stripe==14.1.0
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tnefparse==1.4.0
tokenizers==0.22.2
tqdm==4.67.1
twilio==9.10.0
typer==0.19.2
typer-slim==0.21.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.25.0
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==68.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
zopfli==0.4.0
